GlobalIllum attemp 2

+++ COMP371_RaytracerBase/code/src/Output.h
@@ -17,7 +17,7 @@ public:
     bool globalillum;
     bool antialiasing;
     int raysperpixel;
-    int probterminate = 1 / (2 * pi);
+    double probterminate = 1 / (2 * pi);
     bool twosiderender = true;
     int maxbounces = 1;
